W7D4 Entity Repo Validation

+++ musify/src/main/java/com/example/musify/security/JWTUtils.java
@@ -79,9 +79,9 @@ public class JWTUtils {
         return token;
     }
 
-    public Integer getUserId(){
+    public Long getUserId(){
         List<?> userInfo = (List<?>) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-        return (Integer) userInfo.get(0);
+        return (Long) userInfo.get(0);
     }
